omni-life-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
google-auth==2.30.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.3
pydantic-settings==2.3.1
httpx==0.27.0
